Extract SendPendingMessages
