+++ k9mail/src/test/java/com/fsck/k9/controller/MessagingControllerTest.java
@@ -12,22 +12,17 @@ import android.content.Context;
 
 import com.fsck.k9.Account;
 import com.fsck.k9.AccountStats;
-import com.fsck.k9.K9;
 import com.fsck.k9.K9RobolectricTestRunner;
 import com.fsck.k9.Preferences;
 import com.fsck.k9.controller.tasks.SearchResultsLoader;
 import com.fsck.k9.helper.Contacts;
-import com.fsck.k9.mail.AuthenticationFailedException;
-import com.fsck.k9.mail.CertificateValidationException;
 import com.fsck.k9.mail.FetchProfile;
 import com.fsck.k9.mail.FetchProfile.Item;
-import com.fsck.k9.mail.Flag;
 import com.fsck.k9.mail.Folder;
 import com.fsck.k9.mail.Message;
 import com.fsck.k9.mail.MessageRetrievalListener;
 import com.fsck.k9.mail.MessagingException;
 import com.fsck.k9.mail.Store;
-import com.fsck.k9.mail.Transport;
 import com.fsck.k9.mail.TransportProvider;
 import com.fsck.k9.mailstore.LocalFolder;
 import com.fsck.k9.mailstore.LocalMessage;
@@ -40,7 +35,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
-import org.mockito.InOrder;
 import org.mockito.Matchers;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
@@ -57,13 +51,10 @@ import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.atLeast;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.doAnswer;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.inOrder;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 
 
@@ -71,7 +62,6 @@ import static org.mockito.Mockito.when;
 @RunWith(K9RobolectricTestRunner.class)
 public class MessagingControllerTest {
     private static final String FOLDER_NAME = "Folder";
-    private static final String SENT_FOLDER_NAME = "Sent";
     private static final int MAXIMUM_SMALL_MESSAGE_SIZE = 1000;
     private static final String MESSAGE_UID1 = "message-uid1";
 
@@ -90,8 +80,6 @@ public class MessagingControllerTest {
     @Mock
     private LocalFolder localFolder;
     @Mock
-    private LocalFolder sentFolder;
-    @Mock
     private Folder remoteFolder;
     @Mock
     private LocalStore localStore;
@@ -101,8 +89,6 @@ public class MessagingControllerTest {
     private NotificationController notificationController;
     @Mock
     private TransportProvider transportProvider;
-    @Mock
-    private Transport transport;
     @Mock
     private SearchResultsLoader searchResultsLoader;
     @Captor
@@ -118,8 +104,6 @@ public class MessagingControllerTest {
     private Message remoteNewMessage1;
     @Mock
     private LocalMessage localNewMessage1;
-    @Mock
-    private LocalMessage localMessageToSend1;
 
 
     @Before
@@ -247,114 +231,6 @@ public class MessagingControllerTest {
         assertTrue(fetchProfileCaptor.getAllValues().get(2).contains(Item.STRUCTURE));
         assertEquals(1, fetchProfileCaptor.getAllValues().get(2).size());
     }
-
-    @Test
-    public void sendPendingMessagesSynchronous_withNonExistentOutbox_shouldNotStartSync() throws MessagingException {
-        when(account.getOutboxFolderName()).thenReturn(FOLDER_NAME);
-        when(localFolder.exists()).thenReturn(false);
-        controller.addListener(listener);
-
-        controller.sendPendingMessagesSynchronous(account);
-
-        verifyZeroInteractions(listener);
-    }
-
-    @Test
-    public void sendPendingMessagesSynchronous_shouldCallListenerOnStart() throws MessagingException {
-        setupAccountWithMessageToSend();
-
-        controller.sendPendingMessagesSynchronous(account);
-
-        verify(listener).sendPendingMessagesStarted(account);
-    }
-
-    @Test
-    public void sendPendingMessagesSynchronous_shouldSetProgress() throws MessagingException {
-        setupAccountWithMessageToSend();
-
-        controller.sendPendingMessagesSynchronous(account);
-
-        verify(listener).synchronizeMailboxProgress(account, "Sent", 0, 1);
-    }
-
-    @Test
-    public void sendPendingMessagesSynchronous_shouldSendMessageUsingTransport() throws MessagingException {
-        setupAccountWithMessageToSend();
-
-        controller.sendPendingMessagesSynchronous(account);
-
-        verify(transport).sendMessage(localMessageToSend1);
-    }
-
-    @Test
-    public void sendPendingMessagesSynchronous_shouldSetAndRemoveSendInProgressFlag() throws MessagingException {
-        setupAccountWithMessageToSend();
-
-        controller.sendPendingMessagesSynchronous(account);
-
-        InOrder ordering = inOrder(localMessageToSend1, transport);
-        ordering.verify(localMessageToSend1).setFlag(Flag.X_SEND_IN_PROGRESS, true);
-        ordering.verify(transport).sendMessage(localMessageToSend1);
-        ordering.verify(localMessageToSend1).setFlag(Flag.X_SEND_IN_PROGRESS, false);
-    }
-
-    @Test
-    public void sendPendingMessagesSynchronous_shouldMarkSentMessageAsSeen() throws MessagingException {
-        setupAccountWithMessageToSend();
-
-        controller.sendPendingMessagesSynchronous(account);
-
-        verify(localMessageToSend1).setFlag(Flag.SEEN, true);
-    }
-
-    @Test
-    public void sendPendingMessagesSynchronous_whenMessageSentSuccesfully_shouldUpdateProgress() throws MessagingException {
-        setupAccountWithMessageToSend();
-
-        controller.sendPendingMessagesSynchronous(account);
-
-        verify(listener).synchronizeMailboxProgress(account, "Sent", 1, 1);
-    }
-
-    @Test
-    public void sendPendingMessagesSynchronous_shouldUpdateProgress() throws MessagingException {
-        setupAccountWithMessageToSend();
-
-        controller.sendPendingMessagesSynchronous(account);
-
-        verify(listener).synchronizeMailboxProgress(account, "Sent", 1, 1);
-    }
-
-    @Test
-    public void sendPendingMessagesSynchronous_withAuthenticationFailure_shouldNotify() throws MessagingException {
-        setupAccountWithMessageToSend();
-        doThrow(new AuthenticationFailedException("Test")).when(transport).sendMessage(localMessageToSend1);
-
-        controller.sendPendingMessagesSynchronous(account);
-
-        verify(notificationController).showAuthenticationErrorNotification(account, false);
-    }
-
-    @Test
-    public void sendPendingMessagesSynchronous_withCertificateFailure_shouldNotify() throws MessagingException {
-        setupAccountWithMessageToSend();
-        doThrow(new CertificateValidationException("Test")).when(transport).sendMessage(localMessageToSend1);
-
-        controller.sendPendingMessagesSynchronous(account);
-
-        verify(notificationController).showCertificateErrorNotification(account, false);
-    }
-
-    @Test
-    public void sendPendingMessagesSynchronous_shouldCallListenerOnCompletion() throws MessagingException {
-        setupAccountWithMessageToSend();
-
-        controller.sendPendingMessagesSynchronous(account);
-
-        verify(listener).sendPendingMessagesCompleted(account);
-    }
-
-
     @Test
     public void synchronizeMailboxSynchronousLegacy_withOneMessageInRemoteFolder_shouldFinishWithoutError()
             throws Exception {
@@ -565,20 +441,6 @@ public class MessagingControllerTest {
         assertEquals(FetchProfile.Item.STRUCTURE, fetchProfileCaptor.getAllValues().get(2).get(0));
         assertEquals(1, fetchProfileCaptor.getAllValues().get(3).size());
         assertEquals(FetchProfile.Item.BODY_SANE, fetchProfileCaptor.getAllValues().get(3).get(0));
-    }
-
-    private void setupAccountWithMessageToSend() throws MessagingException {
-        when(account.getOutboxFolderName()).thenReturn(FOLDER_NAME);
-        when(account.hasSentFolder()).thenReturn(true);
-        when(account.getSentFolderName()).thenReturn(SENT_FOLDER_NAME);
-        when(localStore.getFolder(SENT_FOLDER_NAME)).thenReturn(sentFolder);
-        when(sentFolder.getDatabaseId()).thenReturn(1L);
-        when(localFolder.exists()).thenReturn(true);
-        when(transportProvider.getTransport(appContext, account)).thenReturn(transport);
-        when(localFolder.getMessages(null)).thenReturn(Collections.singletonList(localMessageToSend1));
-        when(localMessageToSend1.getUid()).thenReturn("localMessageToSend1");
-        when(localMessageToSend1.getHeader(K9.IDENTITY_HEADER)).thenReturn(new String[]{});
-        controller.addListener(listener);
     }
 
     private void respondToFetchEnvelopesWithMessage(final Message message) throws MessagingException {
